Make more simulations go through
The `enable_lax_arithmetic` flag gets rid of the huge number of spurious
goals we were generating. Assuming that arithmetic is two's complement
and proving that the results are correct under that assumption is
sufficient for this case.

Now there are only 5 functions we're planning to verify that don't go
through. Two result in huge memory use and might benefit from loop
invariants, and three need more precise specifications from their
callees.